RAG with Langchain and Chainlit/requirements.txt
chainlit
openai
langchain
python-dotenv
langchain_openai 
pypdf
langchain-community
sentence-transformers
langchain_huggingface
faiss-cpu
langsmith
langchainhub